fix: 56634da - dry run bug and file not closed
Fixes bug introduced by #68 with expecting publication files to be generated even in dry run.

Fixes bug introduced by #68 with file not closed.

Fix 56634da
Fix #68
See #75

+++ academic/editFM.py
@@ -6,24 +6,29 @@ yaml = YAML()
 
 
 class EditableFM:
-    def __init__(self, base_path: Path, delim: str = "---"):
-        self.base_path = Path(base_path)
+    def __init__(self, base_path: Path, delim: str = "---", dry_run: bool = False):
+        self.base_path = base_path
         if delim != "---":
             raise NotImplementedError("Currently, YAML is the only supported front-matter format.")
         self.delim = delim
         self.fm = []
         self.content = []
         self.path = ""
+        self.dry_run = dry_run
 
     def load(self, file: Path):
         self.fm = []
         self.content = []
         self.path = self.base_path / file
+        if self.dry_run and not self.path.exists():
+            self.fm = dict()
+            return
 
-        file = open(self.path, "r").readlines()
+        with self.path.open("r", encoding="utf-8") as f:
+            lines = open(self.path, "r").readlines()
 
         delims_seen = 0
-        for line in file:
+        for line in lines:
             if line.startswith(self.delim):
                 delims_seen += 1
             else:
@@ -37,6 +42,8 @@ class EditableFM:
 
     def dump(self):
         assert self.path, "You need to `.load()` first."
+        if self.dry_run:
+            return
 
         with open(self.path, "w", encoding="utf-8") as f:
             f.write("{}\n".format(self.delim))
